requirements.txt
nba-api==1.4.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
prophet==1.1.5
tensorflow==2.15.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
streamlit==1.29.0
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3
jupyter==1.0.0
pytest==7.4.3